Clean up change in cli and move checks to validate
