+++ src/Directory.Packages.props
@@ -13,7 +13,6 @@
     <PackageVersion Include="Microsoft.Extensions.Configuration.Binder" Version="9.0.8" />
     <PackageVersion Include="System.Text.Json" Version="9.0.8" />
     <!-- Security -->
-    <PackageVersion Include="Microsoft.Playwright" Version="1.60.0" />
     <PackageVersion Include="Microsoft.Identity.Client" Version="4.78.0" />
     <PackageVersion Include="Microsoft.Identity.Client.Broker" Version="4.78.0" />
     <PackageVersion Include="Microsoft.Identity.Client.Extensions.Msal" Version="4.78.0" />
