+++ src/Microsoft.Agents.A365.DevTools.Cli/Microsoft.Agents.A365.DevTools.Cli.csproj
@@ -45,9 +45,6 @@
     <PackageReference Include="Microsoft.Extensions.Logging" />
     <PackageReference Include="Microsoft.Extensions.Logging.Console" />
 
-    <!-- Browser Automation -->
-    <PackageReference Include="Microsoft.Playwright" />
-
     <!-- Azure SDKs -->
     <PackageReference Include="Azure.Identity" />
     <PackageReference Include="Azure.ResourceManager" />
