+++ src/a365.config.example.json
@@ -5,18 +5,13 @@
   "resourceGroup": "your-resource-group-name",
   "location": "westus",
   "environment": "preprod",
-  "agent365ObservabilityMcpOptions": {
-    "baseUrl": "https://your-observability-host",
-    "agentObservabilityId": "11111111-2222-3333-4444-555555555555",
-    "appId": "42edbbd6-cfc9-4637-9068-ebac6df46171"
-  },
   "appServicePlanName": "your-app-service-plan-name",
   "appServicePlanSku": "B1",
   "webAppName": "your-unique-webapp-name",
-  "agentIdentityDisplayName": "Your Agent Identity Display Name",
-  "agentBlueprintDisplayName": "Your Blueprint App Display Name",
+  "agentIdentityDisplayName": "Your Agent Display Name",
+  "agentBlueprintDisplayName": "Your Blueprint Display Name",
   "agentUserPrincipalName": "agentuser@yourdomain.onmicrosoft.com",
-  "agentUserDisplayName": "Your Agent User Account Display Name",
+  "agentUserDisplayName": "Your Agent User Display Name",
   "managerEmail": "manager@yourdomain.onmicrosoft.com",
   "agentUserUsageLocation": "US",
   "deploymentProjectPath": "/path/to/your/agent/project",
